freebsd-relayd: pfe: Make sure that table names aren't too long
relayd will fail to load a configuration if a table name is too long to
fit into the ioctl request structure, but in this case relayd -n would
happily declare the configuration as OK.  table name limits are fairly
small so this limit is easy to hit.

Modify the buffer size used by relayd so that the parser catches this
error.  Improve error messages from the parser so that it's a bit easier
to see what the problem is.

Note that RD_TAG_NAME_SIZE is the same as PF_TAG_NAME_SIZE on both
FreeBSD and OpenBSD.